Encapsulate store state in internal module
